+++ src/Dynatrace.OpenKit/Core/Objects/EventPayloadBuilder.cs
@@ -1,4 +1,21 @@
-﻿using Dynatrace.OpenKit.API;
+﻿//
+// Copyright 2018-2023 Dynatrace LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+using Dynatrace.OpenKit.API;
+using Dynatrace.OpenKit.Core.Util;
 using Dynatrace.OpenKit.Util.Json.Objects;
 using System.Collections.Generic;
 
@@ -59,6 +76,23 @@ namespace Dynatrace.OpenKit.Core.Objects
         public string Build()
         {
             return JsonObjectValue.FromDictionary(attributes).ToString();
+        }
+
+        /// <summary>
+        /// Checks if the attributes contain a non-finite value
+        /// </summary>
+        /// <returns>True if non-finite values is within attributes</returns>
+        internal bool IsEventPayloadContainingNonFiniteValues()
+        {
+            foreach (KeyValuePair<string, JsonValue> entry in attributes)
+            {
+                if(EventPayloadBuilderUtil.IsItemContainingNonFiniteNumericValues(entry.Value))
+                {
+                    return true;
+                }
+            }
+
+            return false;
         }
 
         /// <summary>